requirements.txt
streamlit
pydantic
google-generativeai
python-dotenv
pytest
docxtpl
python-docx
